extract plugins init codes to another method
